FaceLockAutomationRPI/requirements.txt
applicationinsights==0.11.7
asn1crypto==0.24.0
azure-common==1.1.16
azure-iothub-device-client==1.4.3
azure-nspkg==3.0.2
azure-storage-blob==1.3.1
azure-storage-common==1.3.0
azure-storage-file==1.3.1
azure-storage-nspkg==3.0.0
azure-storage-queue==1.3.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.3.1
dlib==19.16.0
idna==2.7
imutils==0.5.1
numpy==1.15.1
picamera==1.13
Pillow==5.3.0
pipenv==2018.11.26
pycparser==2.19
python-dateutil==2.7.3
requests==2.19.1
RPi.GPIO==0.6.3
six==1.11.0
urllib3==1.23
virtualenv==16.2.0
virtualenv-clone==0.4.0
